+++ hyppo/ksample/mmd.py
@@ -25,7 +25,7 @@ class MMD(KSampleTest):
     :math:`\{ u_i \stackrel{iid}{\sim} F_U,\ i = 1, ..., n \}` and
     :math:`\{ v_j \stackrel{iid}{\sim} F_V,\ j = 1, ..., m \}` as two groups
     of samples deriving from different distributions with the same
-    dimensionality. If :math:`k(\cdot, \cdot)` is a kernel metric (i.e. gaussian)
+    dimensionality. If :math:`k(\cdot, \cdot)` is a kernel metric (i.e. Gaussian)
     then,
 
     .. math::